# back/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AixProject API Documentation",
    "version": "1.0.0",
    "description": "API documentation for AixProject management platform"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "code": {
            "type": "number"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/admin/api-log": {
      "get": {
        "summary": "获取API日志列表",
        "tags": [
          "ApiLog"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "pageNum",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "username",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "method",
            "schema": {
              "type": "string",
              "enum": [
                "GET",
                "POST",
                "PUT",
                "DELETE",
                "PATCH"
              ]
            }
          },
          {
            "in": "query",
            "name": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "delete": {
        "summary": "删除API日志",
        "tags": [
          "ApiLog"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "要删除的日志ID数组，不传则清空所有日志"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "删除成功"
          }
        }
      }
    },
    "/admin/api-log/statistics": {
      "get": {
        "summary": "获取API日志统计",
        "tags": [
          "ApiLog"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/admin/audit": {
      "get": {
        "summary": "获取审计日志列表",
        "tags": [
          "Audit"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "pageNum",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "username",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "method",
            "schema": {
              "type": "string",
              "enum": [
                "GET",
                "POST",
                "PUT",
                "DELETE",
                "PATCH"
              ]
            }
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/admin/audit/{id}/rollback": {
      "post": {
        "summary": "回滚审计日志操作",
        "tags": [
          "Audit"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "审计日志ID"
          }
        ],
        "responses": {
          "200": {
            "description": "回滚成功"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/admin/auth/login": {
      "post": {
        "summary": "用户登录",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "用户名"
                  },
                  "password": {
                    "type": "string",
                    "description": "密码"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功"
          },
          "401": {
            "description": "用户名或密码错误"
          }
        }
      }
    },
    "/admin/auth/refresh": {
      "post": {
        "summary": "刷新Token",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "refreshToken"
                ],
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "description": "刷新令牌"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "刷新成功"
          },
          "401": {
            "description": "刷新令牌无效或已过期"
          }
        }
      }
    },
    "/admin/auth/logout": {
      "post": {
        "summary": "用户登出",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "登出成功"
          }
        }
      }
    },
    "/admin/auth/permissions": {
      "get": {
        "summary": "获取用户权限",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功"
          }
        }
      }
    },
    "/admin/auth/register": {
      "post": {
        "summary": "用户注册（开发测试用）",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password",
                  "name"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "用户名"
                  },
                  "password": {
                    "type": "string",
                    "description": "密码"
                  },
                  "name": {
                    "type": "string",
                    "description": "姓名"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "注册成功"
          }
        }
      }
    },
    "/admin/{entity}": {
      "get": {
        "summary": "获取实体列表（分页）",
        "tags": [
          "CRUD"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "entity",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "实体名称 (sys用户, sys角色, etc.)"
          },
          {
            "in": "query",
            "name": "pageNum",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "summary": "创建实体",
        "tags": [
          "CRUD"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "entity",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功"
          }
        }
      }
    },
    "/admin/{entity}/{id}": {
      "put": {
        "summary": "更新实体",
        "tags": [
          "CRUD"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "entity",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功"
          }
        }
      },
      "delete": {
        "summary": "删除实体",
        "tags": [
          "CRUD"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "entity",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        }
      }
    },
    "/admin/{entity}/batch-delete": {
      "post": {
        "summary": "批量删除实体",
        "tags": [
          "CRUD"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "entity",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "删除成功"
          }
        }
      }
    },
    "/admin/i18n": {
      "get": {
        "summary": "获取所有国际化内容",
        "tags": [
          "I18n"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "保存国际化内容",
        "tags": [
          "I18n"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "国际化内容对象，键为语言代码，值为翻译对象"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "保存成功"
          }
        }
      }
    },
    "/admin/scheduler/task/runs/{taskId}": {
      "get": {
        "summary": "获取定时任务执行记录",
        "tags": [
          "Scheduler"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/admin/scheduler/task/run/{id}": {
      "post": {
        "summary": "手动触发定时任务",
        "tags": [
          "Scheduler"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "任务已触发"
          }
        }
      }
    },
    "/admin/scheduler/task/stop/{code}": {
      "post": {
        "summary": "停止定时任务",
        "tags": [
          "Scheduler"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "任务已停止"
          }
        }
      }
    },
    "/admin/scheduler/task/list": {
      "get": {
        "summary": "获取所有定时任务",
        "tags": [
          "Scheduler"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/common": {
      "get": {
        "summary": "公共模块信息",
        "tags": [
          "Common"
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/common/time": {
      "get": {
        "summary": "获取服务器时间",
        "tags": [
          "Common"
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/common/test-echo": {
      "get": {
        "summary": "GET测试接口（定时任务调用）",
        "tags": [
          "Common"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "param",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "summary": "POST测试接口（定时任务调用）",
        "tags": [
          "Common"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/common/user-count": {
      "get": {
        "summary": "获取用户数量（测试用）",
        "tags": [
          "Common"
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/common/system-info": {
      "get": {
        "summary": "获取系统信息配置",
        "tags": [
          "Common"
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/common/maintenance-status": {
      "get": {
        "summary": "获取维护模式状态",
        "tags": [
          "Common"
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/common/copyright-info": {
      "get": {
        "summary": "获取版权信息",
        "tags": [
          "Common"
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/common/git-logs": {
      "get": {
        "summary": "获取Git提交记录",
        "tags": [
          "Common"
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    }
  },
  "tags": []
}